A metabolomic data preprocessing R scripts
The latest version of R scripts and *.sh files to data preprocessing for
monitoring metabolomic changes in multifactorial experiments: a study of
barley (Hordeum vulgare) leaves under drought stress. Scripts were run
for 1632 biological samples recorded at 2 selected wavelength in 13201
time points (from 2 to 13 min) within a reasonably short time (in a
couple of days on a computing cluster with 6176 CPU cores and the total
computing power of 54.27 TFLOPS).